+++ packages/angular/test/base/src/app/lazy/modal-dynamic-wrapper/index.ts
@@ -1,0 +1,2 @@
+export * from './modal-dynamic-wrapper.component';
+export * from './modal-dynamic-wrapper.module';